feat(plugin-environment-variables): expose $env on the flow-engine context in v1 too
Extract the v2 plugin's `defineProperty('$env')` registration into a
framework-agnostic `registerEnvProperty(context, apiClient, t)` helper, and call
it from the v1 client entry as well (via the allowed v1 -> v2 import). This puts
`$env` on the v1 flow-engine context — the same mechanism v2 already uses — so v2
UI rendered in the v1 runtime (which mounts detached from the v1 React tree and
can't reach the React-context `addGlobalVar('$env')`) can resolve `$env` from
`flowEngine.context.getPropertyMetaTree()`. The list is fetched once and cached by
the flow-engine context. The legacy v1 `addGlobalVar` + provider path is kept
untouched.
